Revert "Change NewDocs to int64"
This reverts commit eb7f2e5fd520c2ad05a310a5777486e808abafc3.

+++ src/dbnode/storage/stats/query_stats_test.go
@@ -85,7 +85,7 @@ func TestPeriodicallyResetRecentDocs(t *testing.T) {
 	verifyStats(t, tracker,  0, 0)
 }
 
-func verifyStats(t *testing.T, tracker *testQueryStatsTracker, expectedNew int64, expectedRecent int64) {
+func verifyStats(t *testing.T, tracker *testQueryStatsTracker, expectedNew int, expectedRecent int64) {
 	values := tracker.StatsValues()
 	assert.Equal(t, expectedNew, values.NewDocs)
 	assert.Equal(t, expectedRecent, values.RecentDocs)
